+++ pcs/cli/common/errors.py
@@ -5,6 +5,12 @@ from __future__ import (
     unicode_literals,
 )
 
+
+ERR_NODE_LIST_AND_ALL_MUTUALLY_EXCLUSIVE = (
+    "Cannot specify both --all and a list of nodes."
+)
+
+
 class CmdLineInputError(Exception):
     """
     Exception express that user entered incorrect commad in command line.