+++ CHANGELOG.md
@@ -32,6 +32,8 @@
 - Fixed parsing commandline options ([rhbz#1404229])
 - Pcs does not exit with a false error message anymore when pcsd-cli.rb outputs
   to stderr ([ghissue#124])
+- Pcs now exits with an error when both `--all` and a list of nodes is specified
+  in the `pcs cluster start | stop | enable | disable` commands ([rhbz#1339355])
 
 [ghissue#124]: https://github.com/ClusterLabs/pcs/issues/124
 [ghissue#125]: https://github.com/ClusterLabs/pcs/issues/125
@@ -40,6 +42,7 @@
 [rhbz#1261116]: https://bugzilla.redhat.com/show_bug.cgi?id=1261116
 [rhbz#1284404]: https://bugzilla.redhat.com/show_bug.cgi?id=1284404
 [rhbz#1315992]: https://bugzilla.redhat.com/show_bug.cgi?id=1315992
+[rhbz#1339355]: https://bugzilla.redhat.com/show_bug.cgi?id=1339355
 [rhbz#1389443]: https://bugzilla.redhat.com/show_bug.cgi?id=1389443
 [rhbz#1389501]: https://bugzilla.redhat.com/show_bug.cgi?id=1389501
 [rhbz#1389941]: https://bugzilla.redhat.com/show_bug.cgi?id=1389941